Get things split up most of the way.
Still need to move GPU binding management to the GlyphCache.
Will clean up the import mess once everything gets moved into
the right project.
