+++ apps/show-mm/Cargo.toml
@@ -16,6 +16,7 @@ absolute_unit = { path = "../../libs/nitrogen/sim/absolute_unit" }
 atmosphere = { path = "../../libs/render-wgpu/buffer/atmosphere" }
 camera = { path = "../../libs/nitrogen/sim/camera" }
 command = { path = "../../libs/nitrogen/system/command" }
+fnt = { path = "../../libs/fnt" }
 frame_graph = { path = "../../libs/nitrogen/system/frame_graph" }
 fullscreen = { path = "../../libs/render-wgpu/buffer/fullscreen" }
 galaxy = { path = "../../libs/galaxy" }
