+++ apps/show-sh/Cargo.toml
@@ -16,11 +16,13 @@ absolute_unit = { path = "../../libs/nitrogen/sim/absolute_unit" }
 atmosphere = { path = "../../libs/render-wgpu/buffer/atmosphere" }
 camera = { path = "../../libs/nitrogen/sim/camera" }
 command = { path = "../../libs/nitrogen/system/command" }
+fnt = { path = "../../libs/fnt" }
 frame_graph = { path = "../../libs/nitrogen/system/frame_graph" }
 fullscreen = { path = "../../libs/render-wgpu/buffer/fullscreen" }
 geodesy = { path = "../../libs/nitrogen/sim/geodesy" }
 galaxy = { path = "../../libs/galaxy" }
 global_data = { path = "../../libs/render-wgpu/buffer/global_data" }
+glyph_cache = { path = "../../libs/render-wgpu/buffer/glyph_cache" }
 gpu = { path = "../../libs/nitrogen/system/gpu" }
 input = { path = "../../libs/nitrogen/system/input" }
 lib = { path = "../../libs/lib" }
