+++ apps/show-sky/src/main.rs
@@ -24,7 +24,6 @@ use geodesy::{GeoSurface, Graticule, Target};
 use global_data::GlobalParametersBuffer;
 use gpu::GPU;
 use input::InputSystem;
-use lib::Library;
 use log::trace;
 use nalgebra::convert;
 use orrery::Orrery;
@@ -36,7 +35,7 @@ use stars::StarsBuffer;
 use std::time::Instant;
 use terrain::TerrainRenderPass;
 use terrain_geo::{CpuDetailLevel, GpuDetailLevel, TerrainGeoBuffer};
-use text_layout::{Font, TextAnchorH, TextAnchorV, TextLayoutBuffer, TextPositionH, TextPositionV};
+use text_layout::{TextAnchorH, TextAnchorV, TextLayoutBuffer, TextPositionH, TextPositionV};
 
 make_frame_graph!(
     FrameGraph {
@@ -59,9 +58,6 @@ make_frame_graph!(
 
 fn main() -> Fallible<()> {
     TermLogger::init(LevelFilter::Warn, Config::default())?;
-
-    use std::sync::Arc;
-    let lib = Arc::new(Box::new(Library::empty()?));
 
     let system_bindings = Bindings::new("map")
         .bind("+target_up", "Up")?
@@ -87,7 +83,7 @@ fn main() -> Fallible<()> {
     let globals_buffer = GlobalParametersBuffer::new(gpu.device())?;
     let stars_buffer = StarsBuffer::new(&gpu)?;
     let terrain_geo_buffer = TerrainGeoBuffer::new(cpu_detail, gpu_detail, &mut gpu)?;
-    let text_layout_buffer = TextLayoutBuffer::new(&lib, &mut gpu)?;
+    let text_layout_buffer = TextLayoutBuffer::new(&mut gpu)?;
 
     let mut frame_graph = FrameGraph::new(
         &mut gpu,
@@ -102,7 +98,7 @@ fn main() -> Fallible<()> {
 
     let fps_handle = text_layout_buffer
         .borrow_mut()
-        .add_screen_text(Font::QUANTICO, "", &gpu)?
+        .add_screen_text("", "", &gpu)?
         .with_color(&[1f32, 0f32, 0f32, 1f32])
         .with_horizontal_position(TextPositionH::Left)
         .with_horizontal_anchor(TextAnchorH::Left)
