+++ libs/render-wgpu/buffer/text_layout/src/layout_vertex.rs
@@ -1,0 +1,59 @@
+// This file is part of OpenFA.
+//
+// OpenFA is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// OpenFA is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with OpenFA.  If not, see <http://www.gnu.org/licenses/>.
+use memoffset::offset_of;
+use std::mem;
+use zerocopy::{AsBytes, FromBytes};
+
+#[repr(C)]
+#[derive(AsBytes, FromBytes, Copy, Clone, Debug, Default)]
+pub struct LayoutVertex {
+    pub(crate) position: [f32; 2],
+    pub(crate) tex_coord: [f32; 2],
+}
+
+impl LayoutVertex {
+    #[allow(clippy::unneeded_field_pattern)]
+    pub fn descriptor() -> wgpu::VertexBufferDescriptor<'static> {
+        let tmp = wgpu::VertexBufferDescriptor {
+            stride: mem::size_of::<Self>() as wgpu::BufferAddress,
+            step_mode: wgpu::InputStepMode::Vertex,
+            attributes: &[
+                // position
+                wgpu::VertexAttributeDescriptor {
+                    format: wgpu::VertexFormat::Float2,
+                    offset: 0,
+                    shader_location: 0,
+                },
+                // tex_coord
+                wgpu::VertexAttributeDescriptor {
+                    format: wgpu::VertexFormat::Float2,
+                    offset: 8,
+                    shader_location: 1,
+                },
+            ],
+        };
+
+        assert_eq!(
+            tmp.attributes[0].offset,
+            offset_of!(LayoutVertex, position) as wgpu::BufferAddress
+        );
+        assert_eq!(
+            tmp.attributes[1].offset,
+            offset_of!(LayoutVertex, tex_coord) as wgpu::BufferAddress
+        );
+
+        tmp
+    }
+}
