+++ libs/render-wgpu/buffer/text_layout/src/lib.rs
@@ -12,64 +12,24 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with OpenFA.  If not, see <http://www.gnu.org/licenses/>.
+mod layout;
+mod layout_vertex;
+
+use crate::layout::Layout;
+pub use crate::layout_vertex::LayoutVertex;
+
 use failure::Fallible;
-use fnt::Fnt;
 use frame_graph::FrameStateTracker;
-use glyph_cache::{GlyphCache, GlyphCacheIndex};
+use glyph_cache::{GlyphCache, GlyphCacheIndex, TtfFont};
 use gpu::GPU;
-use lib::Library;
 use log::trace;
-use memoffset::offset_of;
 use std::{cell::RefCell, collections::HashMap, mem, ops::Range, sync::Arc};
 use zerocopy::{AsBytes, FromBytes};
 
 // Fallback for when we have no libs loaded.
 // https://fonts.google.com/specimen/Quantico?selection.family=Quantico
+pub const FALLBACK_FONT_NAME: &'static str = "quantico";
 const QUANTICO_TTF_DATA: &[u8] = include_bytes!("../../../../../assets/font/quantico.ttf");
-
-const SPACE_WIDTH: f32 = 5f32 / 640f32;
-
-#[repr(C)]
-#[derive(AsBytes, FromBytes, Copy, Clone, Debug, Default)]
-pub struct LayoutVertex {
-    position: [f32; 2],
-    tex_coord: [f32; 2],
-}
-
-impl LayoutVertex {
-    #[allow(clippy::unneeded_field_pattern)]
-    pub fn descriptor() -> wgpu::VertexBufferDescriptor<'static> {
-        let tmp = wgpu::VertexBufferDescriptor {
-            stride: mem::size_of::<Self>() as wgpu::BufferAddress,
-            step_mode: wgpu::InputStepMode::Vertex,
-            attributes: &[
-                // position
-                wgpu::VertexAttributeDescriptor {
-                    format: wgpu::VertexFormat::Float2,
-                    offset: 0,
-                    shader_location: 0,
-                },
-                // tex_coord
-                wgpu::VertexAttributeDescriptor {
-                    format: wgpu::VertexFormat::Float2,
-                    offset: 8,
-                    shader_location: 1,
-                },
-            ],
-        };
-
-        assert_eq!(
-            tmp.attributes[0].offset,
-            offset_of!(LayoutVertex, position) as wgpu::BufferAddress
-        );
-        assert_eq!(
-            tmp.attributes[1].offset,
-            offset_of!(LayoutVertex, tex_coord) as wgpu::BufferAddress
-        );
-
-        tmp
-    }
-}
 
 #[derive(Copy, Clone, Debug)]
 pub enum TextAnchorH {
@@ -159,382 +119,24 @@ struct LayoutTextRenderContext {
     index_count: u32,
 }
 
-#[repr(C)]
-#[derive(AsBytes, FromBytes, Copy, Clone, Debug)]
-struct LayoutData {
-    text_layout_position: [f32; 4],
-    text_layout_color: [f32; 4],
-}
-
-// Note that each layout has its own vertex/index buffer and a tiny transform
-// buffer that might get updated every frame. This is costly per layout. However,
-// these are screen text layouts, so there will hopefully never be too many of them
-// if we do end up creating lots, we'll need to do some sort of layout caching.
-pub struct Layout {
-    // The externally exposed handle, for ease of use.
-    layout_handle: LayoutHandle,
-
-    // The font used for rendering this layout.
-    glyph_cache_index: GlyphCacheIndex,
-
-    // Cached per-frame render state.
-    content: String,
-    position_x: TextPositionH,
-    position_y: TextPositionV,
-    anchor_x: TextAnchorH,
-    anchor_y: TextAnchorV,
-    color: [f32; 4],
-
-    // Gpu resources
-    text_render_context: Option<LayoutTextRenderContext>,
-    layout_data_buffer: Arc<Box<wgpu::Buffer>>,
-    bind_group: Arc<Box<wgpu::BindGroup>>,
-}
-
-impl Layout {
-    fn new(
-        layout_handle: LayoutHandle,
-        text: &str,
-        glyph_cache: &GlyphCache,
-        bind_group_layout: &wgpu::BindGroupLayout,
-        gpu: &GPU,
-    ) -> Fallible<Self> {
-        let size = mem::size_of::<LayoutData>() as wgpu::BufferAddress;
-        let layout_data_buffer = Arc::new(Box::new(gpu.device().create_buffer(
-            &wgpu::BufferDescriptor {
-                label: Some("text-layout-data-buffer"),
-                size,
-                usage: wgpu::BufferUsage::all(),
-            },
-        )));
-
-        let bind_group = gpu.device().create_bind_group(&wgpu::BindGroupDescriptor {
-            label: Some("text-layout-bind-group"),
-            layout: &bind_group_layout,
-            bindings: &[wgpu::Binding {
-                binding: 0,
-                resource: wgpu::BindingResource::Buffer {
-                    buffer: &layout_data_buffer,
-                    range: 0..size,
-                },
-            }],
-        });
-
-        let text_render_context = Self::build_text_span(text, &glyph_cache, gpu)?;
-        Ok(Self {
-            layout_handle,
-            glyph_cache_index: glyph_cache.index(),
-
-            content: text.to_owned(),
-            position_x: TextPositionH::Center,
-            position_y: TextPositionV::Center,
-            anchor_x: TextAnchorH::Left,
-            anchor_y: TextAnchorV::Top,
-            color: [1f32, 0f32, 1f32, 1f32],
-
-            text_render_context: Some(text_render_context),
-            layout_data_buffer,
-            bind_group: Arc::new(Box::new(bind_group)),
-        })
-    }
-
-    pub fn with_span(&mut self, span: &str) -> &mut Self {
-        self.set_span(span);
-        self
-    }
-
-    pub fn with_color(&mut self, clr: &[f32; 4]) -> &mut Self {
-        self.set_color(clr);
-        self
-    }
-
-    pub fn with_horizontal_position(&mut self, pos: TextPositionH) -> &mut Self {
-        self.set_horizontal_position(pos);
-        self
-    }
-
-    pub fn with_vertical_position(&mut self, pos: TextPositionV) -> &mut Self {
-        self.set_vertical_position(pos);
-        self
-    }
-
-    pub fn with_horizontal_anchor(&mut self, anchor: TextAnchorH) -> &mut Self {
-        self.set_horizontal_anchor(anchor);
-        self
-    }
-
-    pub fn with_vertical_anchor(&mut self, anchor: TextAnchorV) -> &mut Self {
-        self.set_vertical_anchor(anchor);
-        self
-    }
-
-    pub fn set_horizontal_position(&mut self, pos: TextPositionH) {
-        self.position_x = pos;
-    }
-
-    pub fn set_vertical_position(&mut self, pos: TextPositionV) {
-        self.position_y = pos;
-    }
-
-    pub fn set_horizontal_anchor(&mut self, anchor: TextAnchorH) {
-        self.anchor_x = anchor;
-    }
-
-    pub fn set_vertical_anchor(&mut self, anchor: TextAnchorV) {
-        self.anchor_y = anchor;
-    }
-
-    pub fn set_color(&mut self, color: &[f32; 4]) {
-        self.color = *color;
-    }
-
-    pub fn set_span(&mut self, text: &str) {
-        self.text_render_context = None;
-        self.content = text.to_owned();
-    }
-
-    pub fn handle(&self) -> LayoutHandle {
-        self.layout_handle
-    }
-
-    fn build_text_span(
-        text: &str,
-        glyph_cache: &GlyphCache,
-        gpu: &GPU,
-    ) -> Fallible<LayoutTextRenderContext> {
-        let mut verts = Vec::new();
-        let mut indices: Vec<u32> = Vec::new();
-
-        let mut offset = 0f32;
-        let mut prior = None;
-        for mut c in text.chars() {
-            if c == ' ' {
-                offset += SPACE_WIDTH;
-                continue;
-            }
-
-            if !glyph_cache.can_render_char(c) {
-                c = '?';
-            }
-
-            let frame = glyph_cache.frame_for(c);
-            let kerning = if let Some(p) = prior {
-                glyph_cache.pair_kerning(p, c)
-            } else {
-                0f32
-            };
-            prior = Some(c);
-
-            // Always do layout from 0-> and let our transform put us in the right spot.
-            let x0 = offset + frame.left_side_bearing + kerning;
-            let x1 = offset + frame.advance_width;
-            let y0 = 0f32;
-            let y1 = glyph_cache.render_height();
-
-            let base = verts.len() as u32;
-            verts.push(LayoutVertex {
-                position: [x0, y0],
-                tex_coord: [frame.s0, 0f32],
-            });
-            verts.push(LayoutVertex {
-                position: [x0, y1],
-                tex_coord: [frame.s0, 1f32],
-            });
-            verts.push(LayoutVertex {
-                position: [x1, y0],
-                tex_coord: [frame.s1, 0f32],
-            });
-            verts.push(LayoutVertex {
-                position: [x1, y1],
-                tex_coord: [frame.s1, 1f32],
-            });
-
-            indices.push(base);
-            indices.push(base + 1u32);
-            indices.push(base + 3u32);
-            indices.push(base);
-            indices.push(base + 3u32);
-            indices.push(base + 2u32);
-
-            offset += frame.advance_width;
-        }
-
-        // Create a degenerate triangle if there was nothing to render, because we cannot
-        // push an empty buffer, but still want something to hold here.
-        if verts.is_empty() {
-            for i in 0..3 {
-                verts.push(LayoutVertex {
-                    position: [0f32, 0f32],
-                    tex_coord: [0f32, 0f32],
-                });
-                indices.push(i);
-            }
-        }
-
-        let vertex_buffer = gpu.push_slice(
-            "text-layout-vertex-buffer",
-            &verts,
-            wgpu::BufferUsage::all(),
-        );
-        let index_buffer = gpu.push_slice(
-            "text-layout-index-buffer",
-            &indices,
-            wgpu::BufferUsage::all(),
-        );
-
-        Ok(LayoutTextRenderContext {
-            render_width: offset,
-            vertex_buffer: Arc::new(Box::new(vertex_buffer)),
-            index_buffer: Arc::new(Box::new(index_buffer)),
-            index_count: indices.len() as u32,
-        })
-    }
-
-    fn make_upload_buffer(
-        &mut self,
-        glyph_cache: &GlyphCache,
-        gpu: &GPU,
-        tracker: &mut FrameStateTracker,
-    ) -> Fallible<()> {
-        if self.text_render_context.is_none() {
-            self.text_render_context =
-                Some(Layout::build_text_span(&self.content, &glyph_cache, gpu)?);
-        }
-
-        let x = self.position_x.to_vulkan();
-        let y = self.position_y.to_vulkan();
-
-        let dx = match self.anchor_x {
-            TextAnchorH::Left => 0f32,
-            TextAnchorH::Right => -self.text_render_context.as_ref().unwrap().render_width,
-            TextAnchorH::Center => -self.text_render_context.as_ref().unwrap().render_width / 2f32,
-        };
-
-        let dy = match self.anchor_y {
-            TextAnchorV::Top => 0f32,
-            TextAnchorV::Bottom => -glyph_cache.render_height(),
-            TextAnchorV::Center => -glyph_cache.render_height() / 2f32,
-        };
-
-        let buffer = gpu.push_slice(
-            "text-layout-upload-buffer",
-            &[LayoutData {
-                text_layout_position: [x + dx, y + dy, 0f32, 0f32],
-                text_layout_color: self.color,
-            }],
-            wgpu::BufferUsage::all(),
-        );
-        tracker.upload(
-            buffer,
-            self.layout_data_buffer.clone(),
-            mem::size_of::<LayoutData>(),
-        );
-
-        Ok(())
-    }
-
-    pub fn vertex_buffer(&self) -> &wgpu::Buffer {
-        &self.text_render_context.as_ref().unwrap().vertex_buffer
-    }
-
-    pub fn index_buffer(&self) -> &wgpu::Buffer {
-        &self.text_render_context.as_ref().unwrap().index_buffer
-    }
-
-    pub fn index_range(&self) -> Range<u32> {
-        0u32..self.text_render_context.as_ref().unwrap().index_count
-    }
-
-    pub fn bind_group(&self) -> &wgpu::BindGroup {
-        &self.bind_group
-    }
-}
-
-const _FONT_FILES: [&str; 12] = [
-    "4X12.FNT",
-    "4X6.FNT",
-    "HUD00.FNT",
-    "HUD01.FNT",
-    "HUD11.FNT",
-    "HUDSYM00.FNT",
-    "HUDSYM01.FNT",
-    "HUDSYM11.FNT",
-    "MAPFONT.FNT",
-    "WIN00.FNT",
-    "WIN01.FNT",
-    "WIN11.FNT",
-];
-
-const _FONT_BACKGROUNDS: [&str; 21] = [
-    "ARMFONT.PIC",
-    "BODYFONT.PIC",
-    "BOLDFONT.PIC",
-    "FNTWPNB.PIC",
-    "FNTWPNY.PIC",
-    "FONT4X6.PIC",
-    "FONTACD.PIC",
-    "FONTACT.PIC",
-    "FONTDFD.PIC",
-    "FONTDFT.PIC",
-    "HEADFONT.PIC",
-    "LRGFONT.PIC",
-    "MAPFONT.PIC",
-    "MENUFONT.PIC",
-    "MFONT320.PIC",
-    "MPFONT.PIC",
-    "PANELFNT.PIC",
-    "PANLFNT2.PIC",
-    "SMLFONT.PIC",
-    "VIDEOFNT.PIC",
-    "WHEELFNT.PIC",
-];
-
-#[derive(Debug, Copy, Clone, Eq, PartialEq, Hash)]
-pub enum Font {
-    HUD11,
-    QUANTICO,
-}
+pub type FontName = String;
 
 pub struct TextLayoutBuffer {
-    glyph_cache_map: HashMap<Font, GlyphCacheIndex>,
+    glyph_cache_map: HashMap<FontName, GlyphCacheIndex>,
     glyph_caches: Vec<GlyphCache>,
-    layout_map: HashMap<Font, Vec<LayoutHandle>>,
+    layout_map: HashMap<FontName, Vec<LayoutHandle>>,
     layouts: Vec<Layout>,
     glyph_bind_group_layout: wgpu::BindGroupLayout,
     layout_bind_group_layout: wgpu::BindGroupLayout,
 }
 
 impl TextLayoutBuffer {
-    pub fn new(lib: &Library, gpu: &mut GPU) -> Fallible<Arc<RefCell<Self>>> {
+    pub fn new(gpu: &mut GPU) -> Fallible<Arc<RefCell<Self>>> {
         trace!("LayoutBuffer::new");
 
         let glyph_bind_group_layout = GlyphCache::create_bind_group_layout(gpu.device());
         let mut glyph_caches = Vec::new();
         let mut glyph_cache_map = HashMap::new();
-
-        // Cache all standard fonts on the GPU.
-        for (name, filename) in &[(Font::HUD11, "HUD11.FNT")] {
-            if let Ok(data) = lib.load(filename) {
-                let index = GlyphCacheIndex::new(glyph_caches.len());
-                glyph_cache_map.insert(*name, index);
-                glyph_caches.push(GlyphCache::new_transparent_fnt(
-                    &Fnt::from_bytes("", "", &data)?,
-                    index,
-                    &glyph_bind_group_layout,
-                    gpu,
-                )?);
-            }
-        }
-
-        // Add fallback font.
-        let index = GlyphCacheIndex::new(glyph_caches.len());
-        glyph_cache_map.insert(Font::QUANTICO, index);
-        glyph_caches.push(GlyphCache::new_ttf(
-            &QUANTICO_TTF_DATA,
-            index,
-            &glyph_bind_group_layout,
-            gpu,
-        )?);
 
         let layout_bind_group_layout =
             gpu.device()
@@ -550,6 +152,14 @@ impl TextLayoutBuffer {
                     }],
                 });
 
+        // Add fallback font.
+        let index = GlyphCacheIndex::new(glyph_caches.len());
+        glyph_cache_map.insert(FALLBACK_FONT_NAME.to_owned(), index);
+        glyph_caches.push(GlyphCache::new(
+            index,
+            TtfFont::new(&QUANTICO_TTF_DATA, &glyph_bind_group_layout, gpu)?,
+        ));
+
         Ok(Arc::new(RefCell::new(Self {
             glyph_cache_map,
             glyph_caches,
@@ -558,6 +168,24 @@ impl TextLayoutBuffer {
             glyph_bind_group_layout,
             layout_bind_group_layout,
         })))
+    }
+
+    pub fn load_font(&mut self, _font_name: FontName, _glyphs: GlyphCache) {
+        // Cache all standard fonts on the GPU.
+        /*
+        for (name, filename) in &[(Font::HUD11, "HUD11.FNT")] {
+            if let Ok(data) = lib.load(filename) {
+                let index = GlyphCacheIndex::new(glyph_caches.len());
+                glyph_cache_map.insert(*name, index);
+                glyph_caches.push(GlyphCache::new_transparent_fnt(
+                    &Fnt::from_bytes("", "", &data)?,
+                    index,
+                    &glyph_bind_group_layout,
+                    gpu,
+                )?);
+            }
+        }
+         */
     }
 
     pub fn glyph_bind_group_layout(&self) -> &wgpu::BindGroupLayout {
@@ -572,7 +200,7 @@ impl TextLayoutBuffer {
         &self.layouts
     }
 
-    pub fn layouts_by_font(&self) -> &HashMap<Font, Vec<LayoutHandle>> {
+    pub fn layouts_by_font(&self) -> &HashMap<FontName, Vec<LayoutHandle>> {
         &self.layout_map
     }
 
@@ -584,14 +212,22 @@ impl TextLayoutBuffer {
         &mut self.layouts[handle.0]
     }
 
-    pub fn glyph_cache(&self, font: Font) -> &GlyphCache {
-        &self.glyph_caches[self.glyph_cache_map[&font].index()]
+    pub fn glyph_cache(&self, font_name: &str) -> &GlyphCache {
+        if let Some(id) = self.glyph_cache_map.get(font_name) {
+            return &self.glyph_caches[id.index()];
+        }
+        &self.glyph_caches[self.glyph_cache_map[FALLBACK_FONT_NAME].index()]
     }
 
     // pub fn layout_mut(&mut self, handle: LayoutHandle) -> &mut Layout {}
 
-    pub fn add_screen_text(&mut self, font: Font, text: &str, gpu: &GPU) -> Fallible<&mut Layout> {
-        let glyph_cache = self.glyph_cache(font);
+    pub fn add_screen_text(
+        &mut self,
+        font_name: &str,
+        text: &str,
+        gpu: &GPU,
+    ) -> Fallible<&mut Layout> {
+        let glyph_cache = self.glyph_cache(font_name);
         let handle = LayoutHandle(self.layouts.len());
         let layout = Layout::new(
             handle,
@@ -602,7 +238,7 @@ impl TextLayoutBuffer {
         )?;
         self.layouts.push(layout);
         self.layout_map
-            .entry(font)
+            .entry(font_name.to_owned())
             .and_modify(|e| e.push(handle))
             .or_insert_with(|| vec![handle]);
         Ok(self.layout_mut(handle))
@@ -615,7 +251,7 @@ impl TextLayoutBuffer {
     ) -> Fallible<()> {
         for layout in self.layouts.iter_mut() {
             layout.make_upload_buffer(
-                &self.glyph_caches[layout.glyph_cache_index.index()],
+                &self.glyph_caches[layout.glyph_cache_index()],
                 gpu,
                 tracker,
             )?;
@@ -639,13 +275,14 @@ mod test {
             "USNF", "MF", "ATF", "ATFNATO", "ATFGOLD", "USNF97", "FA",
         ])?;
         for (game, lib) in omni.libraries() {
+            // FIXME: Load all fonts for the game.
             println!("At: {}", game);
 
-            let layout_buffer = TextLayoutBuffer::new(&lib, &mut gpu)?;
+            let layout_buffer = TextLayoutBuffer::new(&mut gpu)?;
 
             layout_buffer
                 .borrow_mut()
-                .add_screen_text(Font::HUD11, "Top Left (r)", &gpu)?
+                .add_screen_text("quantico", "Top Left (r)", &gpu)?
                 .with_color(&[1f32, 0f32, 0f32, 1f32])
                 .with_horizontal_position(TextPositionH::Left)
                 .with_horizontal_anchor(TextAnchorH::Left)
@@ -655,7 +292,7 @@ mod test {
 
             layout_buffer
                 .borrow_mut()
-                .add_screen_text(Font::HUD11, "Top Right (b)", &gpu)?
+                .add_screen_text("quantico", "Top Right (b)", &gpu)?
                 .with_color(&[0f32, 0f32, 1f32, 1f32])
                 .with_horizontal_position(TextPositionH::Right)
                 .with_horizontal_anchor(TextAnchorH::Right)
@@ -664,7 +301,7 @@ mod test {
 
             layout_buffer
                 .borrow_mut()
-                .add_screen_text(Font::HUD11, "Bottom Left (w)", &gpu)?
+                .add_screen_text("quantico", "Bottom Left (w)", &gpu)?
                 .with_color(&[1f32, 1f32, 1f32, 1f32])
                 .with_horizontal_position(TextPositionH::Left)
                 .with_horizontal_anchor(TextAnchorH::Left)
@@ -673,7 +310,7 @@ mod test {
 
             layout_buffer
                 .borrow_mut()
-                .add_screen_text(Font::HUD11, "Bottom Right (m)", &gpu)?
+                .add_screen_text("quantico", "Bottom Right (m)", &gpu)?
                 .with_color(&[1f32, 0f32, 1f32, 1f32])
                 .with_horizontal_position(TextPositionH::Right)
                 .with_horizontal_anchor(TextAnchorH::Right)
@@ -682,7 +319,7 @@ mod test {
 
             let handle_clr = layout_buffer
                 .borrow_mut()
-                .add_screen_text(Font::HUD11, "", &gpu)?
+                .add_screen_text("quantico", "", &gpu)?
                 .with_span("THR: AFT  1.0G   2462   LCOS   740 M61")
                 .with_color(&[1f32, 0f32, 0f32, 1f32])
                 .with_horizontal_position(TextPositionH::Center)
@@ -693,7 +330,7 @@ mod test {
 
             let handle_fin = layout_buffer
                 .borrow_mut()
-                .add_screen_text(Font::HUD11, "DONE: 0%", &gpu)?
+                .add_screen_text("quantico", "DONE: 0%", &gpu)?
                 .with_color(&[0f32, 1f32, 0f32, 1f32])
                 .with_horizontal_position(TextPositionH::Center)
                 .with_horizontal_anchor(TextAnchorH::Center)
@@ -730,12 +367,11 @@ mod test {
         let input = InputSystem::new(vec![])?;
         let mut gpu = GPU::new(&input, Default::default())?;
 
-        let lib = Arc::new(Box::new(Library::empty()?));
-        let layout_buffer = TextLayoutBuffer::new(&lib, &mut gpu)?;
+        let layout_buffer = TextLayoutBuffer::new(&mut gpu)?;
 
         layout_buffer
             .borrow_mut()
-            .add_screen_text(Font::QUANTICO, "Top Left (r)", &gpu)?
+            .add_screen_text("quantico", "Top Left (r)", &gpu)?
             .with_color(&[1f32, 0f32, 0f32, 1f32])
             .with_horizontal_position(TextPositionH::Left)
             .with_horizontal_anchor(TextAnchorH::Left)
@@ -744,7 +380,7 @@ mod test {
 
         layout_buffer
             .borrow_mut()
-            .add_screen_text(Font::QUANTICO, "Top Right (b)", &gpu)?
+            .add_screen_text("quantico", "Top Right (b)", &gpu)?
             .with_color(&[0f32, 0f32, 1f32, 1f32])
             .with_horizontal_position(TextPositionH::Right)
             .with_horizontal_anchor(TextAnchorH::Right)
@@ -753,7 +389,7 @@ mod test {
 
         layout_buffer
             .borrow_mut()
-            .add_screen_text(Font::QUANTICO, "Bottom Left (w)", &gpu)?
+            .add_screen_text("quantico", "Bottom Left (w)", &gpu)?
             .with_color(&[1f32, 1f32, 1f32, 1f32])
             .with_horizontal_position(TextPositionH::Left)
             .with_horizontal_anchor(TextAnchorH::Left)
@@ -762,7 +398,7 @@ mod test {
 
         layout_buffer
             .borrow_mut()
-            .add_screen_text(Font::QUANTICO, "Bottom Right (m)", &gpu)?
+            .add_screen_text("quantico", "Bottom Right (m)", &gpu)?
             .with_color(&[1f32, 0f32, 1f32, 1f32])
             .with_horizontal_position(TextPositionH::Right)
             .with_horizontal_anchor(TextAnchorH::Right)
@@ -771,7 +407,7 @@ mod test {
 
         let handle_clr = layout_buffer
             .borrow_mut()
-            .add_screen_text(Font::QUANTICO, "", &gpu)?
+            .add_screen_text("quantico", "", &gpu)?
             .with_span("THR: AFT  1.0G   2462   LCOS   740 M61")
             .with_color(&[1f32, 0f32, 0f32, 1f32])
             .with_horizontal_position(TextPositionH::Center)
@@ -782,7 +418,7 @@ mod test {
 
         let handle_fin = layout_buffer
             .borrow_mut()
-            .add_screen_text(Font::QUANTICO, "DONE: 0%", &gpu)?
+            .add_screen_text("quantico", "DONE: 0%", &gpu)?
             .with_color(&[0f32, 1f32, 0f32, 1f32])
             .with_horizontal_position(TextPositionH::Center)
             .with_horizontal_anchor(TextAnchorH::Center)
